refactor(daemon): merge gc_daemon into always-on session daemon (closes #135)
There were two background daemons per user: the centralized session
daemon (`daemon/`) hosting `--detach`/`attach`/`list`/`kill`/`logs`/
repeat jobs, and a separate `gc_daemon` (Phase 1 of #135) owning
`~/.clud/data.redb` for `clud gc *`. The user instinct is always
"there's only one clud daemon, right?" — and PR #151's attempt at
making the session daemon the interactive default left an opening
to fill the "always-on" slot.

This PR folds `gc_daemon` into the session daemon so there is exactly
one background `clud` process per user, auto-spawned from `main.rs` on
every non-`--no-daemon` / non-`--dry-run` invocation. The redb
single-owner invariant (DD-006) survives — only the owning process
identity moved. See DD-012 for full rationale.

Foreground interactive launches still use the direct runner by default
(don't reintroduce the PR #151 attach-pump bug that #152 reverted). The
daemon hosts the centralized PTY path only when explicitly opted in
(`--detach`, `--detachable`, `--experimental-daemon-centralized`,
repeat jobs).

Changes:

* New `daemon/gc_service.rs` — registry worker thread lifted from
  `gc_daemon::run_worker_loop`; same `mpsc::sync_channel(1)` reply
  pattern, same redb single-owner invariant. 4 unit tests carried over.

* `daemon/types.rs` — added `DaemonRequest::Gc { payload: GcOp }` and
  `DaemonResponse::Gc { reply: GcReply }`. New `GcOp` / `GcReply`
  enums + `ListRow` (kept the gc_daemon JSON row shape so
  `clud gc list --json` output is byte-stable for consumers).

* `daemon/server.rs` — spawns the GC registry worker on startup; new
  `dispatch_gc_op` routes `DaemonRequest::Gc` to the worker via a
  per-request sync_channel(1) reply with a 30s timeout.

* `daemon/client.rs` — added fs4 bringup lock (`daemon.lock`,
  issue #138) to `ensure_daemon` so two `clud` startups don't race on
  the TCP bind. Added public `gc_client_{list,purge,reconcile,insert}`
  wrappers for `gc.rs` to call. `ensure_daemon` is now `pub` (callable
  from main.rs which is a separate binary crate).

* `daemon/paths.rs` — daemon state dir defaults to `~/.clud/state`
  (persistent) instead of `$TMP/clud-daemon` (transient). Aligns with
  where `gc_daemon` already kept the redb file.

* `gc.rs` — every `gc_daemon::client_*` call replaced with the
  equivalent `daemon::gc_client_*`. CLI handlers now take a `state_dir`
  parameter resolved once in `run()`.

* `main.rs` — replaced the `gc_daemon::ensure_running()` call site at
  L196 with `daemon::ensure_daemon(&default_state_dir())`. Dropped the
  `__gc-daemon` dispatch at L67 and the
  `!daemon::experimental_enabled(&args)` skip (no longer needed —
  there's only one daemon to spawn).

* `args.rs` — deleted the `InternalGcDaemon` variant and the
  `__gc-daemon` reference in `subcommands`. `--daemon=MODE` flag kept
  as a forward-compat no-op.

* `lib.rs` + `gc_daemon.rs` deleted.

* Docs: rewrote the GC-daemon section of `docs/architecture/gc-and-registry.md`,
  updated `docs/ARCHITECTURE.md`, `docs/architecture/README.md`, and
  `daemon/README.md` to describe the single always-on daemon hosting
  both session ops and the GC registry. Appended DD-012 to
  `DESIGN_DECISIONS.md` recording the merge rationale and noting that
  DD-006's single-owner invariant survives.

Test:
- `soldr cargo fmt --all -- --check` clean
- `soldr cargo clippy --workspace --all-targets -- -D warnings` clean
- `ruff check src tests ci` clean
- `soldr cargo test -p clud --lib` → 567 passed; 0 failed (includes 4
  new gc_service tests + the existing daemon tests)

One-time migration note: users with a running pre-merge `gc_daemon`
process will hit a redb lock conflict on first post-merge run. The old
process idle-shuts after its 30-min window or can be killed manually;
the redb file itself is forward-compatible.
