+++ crates/clud-bin/src/daemon/client.rs
@@ -1,4 +1,4 @@
-use std::fs;
+use std::fs::{self, OpenOptions};
 use std::io::{self, BufRead, BufReader};
 use std::net::{Shutdown, TcpStream};
 use std::path::Path;
@@ -6,22 +6,45 @@ use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::{Duration, Instant};
 
+use fs4::fs_std::FileExt;
+
+use crate::gc::InsertInput;
 use crate::trampoline;
 
 use super::io_helpers::{read_json_file, write_json_line};
-use super::paths::{daemon_info_path, session_snapshot_path, sessions_dir, spec_path, specs_dir};
+use super::paths::{
+    daemon_info_path, daemon_lock_path, session_snapshot_path, sessions_dir, spec_path, specs_dir,
+};
 use super::process_utils::pid_is_alive;
 use super::types::{
-    DaemonInfo, DaemonRequest, DaemonResponse, SessionSnapshot, WorkerClientMessage,
+    DaemonInfo, DaemonRequest, DaemonResponse, GcOp, GcReply, ListRow, SessionSnapshot,
+    WorkerClientMessage,
 };
 
-pub(super) fn ensure_daemon(state_dir: &Path) -> io::Result<()> {
+/// Idempotent best-effort daemon spawn (issue #135). Always called via
+/// `main.rs`; the session daemon is now an always-on background service.
+///
+/// 1. Fast path: read the info file, probe its PID + port; return if up.
+/// 2. Slow path: acquire `<state_dir>/daemon.lock` (issue #138 bringup
+///    serialization), re-probe under the lock, and only spawn `clud
+///    __daemon --state-dir <state_dir>` detached if a sibling didn't
+///    bring the daemon up while we waited. Lock releases when this
+///    function returns.
+///
+/// Visible to `main.rs` (the `clud` binary) so it can call this during
+/// early startup. `pub` rather than `pub(crate)` because the binary is
+/// a separate crate within the package.
+pub fn ensure_daemon(state_dir: &Path) -> io::Result<()> {
     fs::create_dir_all(state_dir)?;
     cleanup_stale_state(state_dir);
-    if let Ok(info) = read_json_file::<DaemonInfo>(&daemon_info_path(state_dir)) {
-        if TcpStream::connect(("127.0.0.1", info.port)).is_ok() {
-            return Ok(());
-        }
+    if probe_existing(state_dir).is_some() {
+        return Ok(());
+    }
+
+    let _bringup_lock = acquire_bringup_lock(state_dir)?;
+    // Re-probe under the lock: a sibling may have spawned while we waited.
+    if probe_existing(state_dir).is_some() {
+        return Ok(());
     }
 
     let args = vec![
@@ -32,9 +55,11 @@ pub(super) fn ensure_daemon(state_dir: &Path) -> io::Result<()> {
     trampoline::spawn_detached_self(&args)?;
 
     let started = Instant::now();
+    let our_pid = std::process::id();
     loop {
-        if let Ok(info) = read_json_file::<DaemonInfo>(&daemon_info_path(state_dir)) {
-            if TcpStream::connect(("127.0.0.1", info.port)).is_ok() {
+        if let Some(info) = probe_existing(state_dir) {
+            // Make sure we didn't read a stale info file from before the spawn.
+            if info.pid != our_pid {
                 return Ok(());
             }
         }
@@ -46,6 +71,30 @@ pub(super) fn ensure_daemon(state_dir: &Path) -> io::Result<()> {
         }
         thread::sleep(Duration::from_millis(25));
     }
+}
+
+fn probe_existing(state_dir: &Path) -> Option<DaemonInfo> {
+    let info = read_json_file::<DaemonInfo>(&daemon_info_path(state_dir)).ok()?;
+    if !pid_is_alive(info.pid) {
+        return None;
+    }
+    if TcpStream::connect(("127.0.0.1", info.port)).is_ok() {
+        Some(info)
+    } else {
+        None
+    }
+}
+
+fn acquire_bringup_lock(state_dir: &Path) -> io::Result<fs::File> {
+    fs::create_dir_all(state_dir)?;
+    let file = OpenOptions::new()
+        .read(true)
+        .write(true)
+        .create(true)
+        .truncate(false)
+        .open(daemon_lock_path(state_dir))?;
+    FileExt::lock_exclusive(&file)?;
+    Ok(file)
 }
 
 pub(super) fn send_daemon_request(
@@ -155,4 +204,93 @@ pub(super) fn cleanup_stale_state(state_dir: &Path) {
 #[allow(dead_code)]
 pub(super) fn remove_spec_file(state_dir: &Path, session_id: &str) {
     let _ = fs::remove_file(spec_path(state_dir, session_id));
+}
+
+// ---------- GC IPC client wrappers (issue #135) ----------
+//
+// Thin convenience layer around `send_daemon_request` for the GC ops the
+// session daemon now serves (replacing the standalone `gc_daemon`
+// process). Auto-spawn the daemon on first use so the CLI works the
+// same way it did against gc_daemon: `clud gc list` from a cold start
+// brings the daemon up.
+
+fn send_gc(state_dir: &Path, op: GcOp) -> io::Result<GcReply> {
+    ensure_daemon(state_dir)?;
+    match send_daemon_request(state_dir, &DaemonRequest::Gc { payload: op })? {
+        DaemonResponse::Gc { reply } => Ok(reply),
+        DaemonResponse::Error { message } => Err(io::Error::other(message)),
+        other => Err(io::Error::new(
+            io::ErrorKind::InvalidData,
+            format!("unexpected daemon response: {other:?}"),
+        )),
+    }
+}
+
+/// `gc.list` — fetch every tracked row.
+pub fn gc_client_list(state_dir: &Path, kind: Option<&str>) -> io::Result<Vec<ListRow>> {
+    match send_gc(
+        state_dir,
+        GcOp::List {
+            kind: kind.map(String::from),
+        },
+    )? {
+        GcReply::ListOk { rows } => Ok(rows),
+        GcReply::Error { message } => Err(io::Error::other(message)),
+        other => Err(io::Error::other(format!("unexpected gc reply: {other:?}"))),
+    }
+}
+
+/// `gc.purge` — purge entries. `duration = None` -> purge all non-live-locked.
+pub fn gc_client_purge(
+    state_dir: &Path,
+    duration: Option<&str>,
+    kind: Option<&str>,
+    dry_run: bool,
+) -> io::Result<(usize, usize)> {
+    match send_gc(
+        state_dir,
+        GcOp::Purge {
+            duration: duration.map(String::from),
+            kind: kind.map(String::from),
+            dry_run,
+        },
+    )? {
+        GcReply::PurgeOk { removed, skipped } => Ok((removed, skipped)),
+        GcReply::Error { message } => Err(io::Error::other(message)),
+        other => Err(io::Error::other(format!("unexpected gc reply: {other:?}"))),
+    }
+}
+
+/// `gc.reconcile` — walk the given repo's `.claude/worktrees/` and insert
+/// new agent-* subdirs. Returns the number of newly-inserted rows.
+pub fn gc_client_reconcile(state_dir: &Path, repo_root: &Path) -> io::Result<usize> {
+    match send_gc(
+        state_dir,
+        GcOp::Reconcile {
+            repo_root: repo_root.to_string_lossy().to_string(),
+        },
+    )? {
+        GcReply::ReconcileOk { inserted } => Ok(inserted),
+        GcReply::Error { message } => Err(io::Error::other(message)),
+        other => Err(io::Error::other(format!("unexpected gc reply: {other:?}"))),
+    }
+}
+
+/// `gc.insert` — insert a single row if not already present.
+pub fn gc_client_insert(state_dir: &Path, input: &InsertInput) -> io::Result<()> {
+    match send_gc(
+        state_dir,
+        GcOp::Insert {
+            kind: input.kind.clone(),
+            path: input.path.clone(),
+            repo_root: input.repo_root.clone(),
+            branch: input.branch.clone(),
+            agent_id: input.agent_id.clone(),
+            created_unix: Some(input.now_unix),
+        },
+    )? {
+        GcReply::InsertOk => Ok(()),
+        GcReply::Error { message } => Err(io::Error::other(message)),
+        other => Err(io::Error::other(format!("unexpected gc reply: {other:?}"))),
+    }
 }
