+++ crates/clud-bin/src/daemon/mod.rs
@@ -2,6 +2,7 @@ mod attach;
 mod client;
 mod commands;
 mod entry;
+mod gc_service;
 mod io_helpers;
 mod keys;
 mod paths;
@@ -12,4 +13,9 @@ mod types;
 mod worker;
 mod worker_shared;
 
+pub use client::{
+    ensure_daemon, gc_client_insert, gc_client_list, gc_client_purge, gc_client_reconcile,
+};
 pub use entry::{experimental_enabled, handle_special_command, run_centralized_session};
+pub use paths::default_state_dir;
+pub use types::{ListRow, ENV_NO_DAEMON};
